AggrHiCExperiment S4 class

The AggrHiCExperiment extends HiCExperiment class

AggrHiCExperiment(
  file,
  resolution = NULL,
  targets,
  flanking_bins = 50,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  bed = NULL,
  BPPARAM = BiocParallel::bpparam()
)

Arguments

file

CoolFile or plain path to a Hi-C contact file

resolution

Resolution to use with the Hi-C contact file

targets

Set of chromosome coordinates for which interaction counts are extracted from the Hi-C contact file, provided as a GRanges object (for diagnoal-centered loci) or as a GInteractions object (for off-diagonal coordinates).

flanking_bins

Number of bins on each flank of the bins containing input targets.

metadata

list of metadata

topologicalFeatures

topologicalFeatures provided as a named SimpleList

pairsFile

Path to an associated .pairs file

bed

Path to regions file generated by HiC-Pro

BPPARAM

BiocParallel parameters

Slots

fileName

Path of Hi-C contact file

resolutions

Resolutions available in the Hi-C contact file.

resolution

Current resolution

interactions

Genomic Interactions extracted from the Hi-C contact file

scores

Available interaction scores.

slices

Available interaction slices.

topologicalFeatures

Topological features associated with the dataset (e.g. loops (\<Pairs\>), borders (\<GRanges\>), viewpoints (\<GRanges\>), etc...)

pairsFile

Path to the .pairs file associated with the Hi-C contact file

metadata

metadata associated with the Hi-C contact file.

Examples

contacts <- full_contacts_yeast()
#> snapshotDate(): 2022-12-16
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
centros <- topologicalFeatures(contacts, 'centromeres')
x <- AggrHiCExperiment(
  file = fileName(contacts), 
  resolution = 1000,
  targets = centros
)
#> Going through preflight checklist...
#> Parsing the entire contact matrice as a sparse matrix...
#> Modeling distance decay...
#> Filtering for contacts within provided targets...
x
#> `AggrHiCExperiment` object over 16 targets 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/4d164d65a1a_7752" 
#> focus: 16 targets 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 1000 
#> interactions: 10201 
#> scores(4): count balanced expected detrended 
#> slices(4): count balanced expected detrended 
#> topologicalFeatures: targets(16) 
#> pairsFile: N/A 
#> metadata(0):
slices(x, 'count')[1:10, 1:10, 1]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    4   43   54   40   41   38   35   16    5    17
#>  [2,]    0   13   61   52   62   42   25   14    9    23
#>  [3,]    0    0   53   72   93   76   47   25    8    29
#>  [4,]    0    0    0    6   67   56   42   20    8    17
#>  [5,]    0    0    0    0   43   68   71   45   25    39
#>  [6,]    0    0    0    0    0   48   70   45   23    61
#>  [7,]    0    0    0    0    0    0   29   31    9    37
#>  [8,]    0    0    0    0    0    0    0    7   17    26
#>  [9,]    0    0    0    0    0    0    0    0   10    38
#> [10,]    0    0    0    0    0    0    0    0    0    33