AggrHiCExperiment S4 class

The AggrHiCExperiment extends HiCExperiment class

AggrHiCExperiment methods.

Usage

AggrHiCExperiment(
  file,
  resolution = NULL,
  targets,
  flankingBins = 50,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(),
  pairsFile = NULL,
  bed = NULL,
  maxDistance = NULL,
  BPPARAM = BiocParallel::bpparam()
)

# S4 method for AggrHiCExperiment,missing
slices(x)

# S4 method for AggrHiCExperiment,character
slices(x, name)

# S4 method for AggrHiCExperiment,numeric
slices(x, name)

# S4 method for AggrHiCExperiment
show(object)

Arguments

file

CoolFile or plain path to a Hi-C contact file

resolution

Resolution to use with the Hi-C contact file

targets

Set of chromosome coordinates for which interaction counts are extracted from the Hi-C contact file, provided as a GRanges object (for diagnoal-centered loci) or as a GInteractions object (for off-diagonal coordinates).

flankingBins

Number of bins on each flank of the bins containing input targets.

metadata

list of metadata

topologicalFeatures

topologicalFeatures provided as a named SimpleList

pairsFile

Path to an associated .pairs file

bed

Path to regions file generated by HiC-Pro

maxDistance

Maximum distance to use when compiling distance decay

BPPARAM

BiocParallel parameters

x, object

A AggrHiCExperiment object.

name

The name/index of slices to extract.

Value

An AggrHiCExperiment object.

Slots

fileName

Path of Hi-C contact file

resolutions

Resolutions available in the Hi-C contact file.

resolution

Current resolution

interactions

Genomic Interactions extracted from the Hi-C contact file

scores

Available interaction scores.

slices

Available interaction slices.

topologicalFeatures

Topological features associated with the dataset (e.g. loops (\<Pairs\>), borders (\<GRanges\>), viewpoints (\<GRanges\>), etc...)

pairsFile

Path to the .pairs file associated with the Hi-C contact file

metadata

metadata associated with the Hi-C contact file.

See also

Examples

fpath <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
#> snapshotDate(): 2023-02-14
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
data(centros_yeast)
x <- AggrHiCExperiment(
  file = fpath, 
  resolution = 8000,
  targets = centros_yeast[c(4, 7)]
)
#> Going through preflight checklist...
#> Parsing the entire contact matrice as a sparse matrix...
#> Modeling distance decay...
#> Filtering for contacts within provided targets...
x
#> `AggrHiCExperiment` object over 2 targets 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/4dd544ed9572_7752" 
#> focus: 2 targets 
#> resolutions(5): 1000 2000 4000 8000 16000
#> active resolution: 8000 
#> interactions: 10201 
#> scores(4): count balanced expected detrended 
#> slices(4): count balanced expected detrended 
#> topologicalFeatures: targets(2) 
#> pairsFile: N/A 
#> metadata(0):
slices(x, 'count')[1:10, 1:10, 1]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 1364 1054  679  470  571  450  230  132  135    83
#>  [2,]    0  832 1430  646  617  432  232  138  142    90
#>  [3,]    0    0 1457 1575  864  603  290  227  219   123
#>  [4,]    0    0    0 1227 1398  634  356  210  218   109
#>  [5,]    0    0    0    0 1285 1544  678  400  361   226
#>  [6,]    0    0    0    0    0 1261 1479  591  649   381
#>  [7,]    0    0    0    0    0    0 1186 1349  885   433
#>  [8,]    0    0    0    0    0    0    0 1252 1561   489
#>  [9,]    0    0    0    0    0    0    0    0 1245  1313
#> [10,]    0    0    0    0    0    0    0    0    0  1356