HiCExperiment S4 class

The HiCExperiment class describes Hi-C contact files imported in R, either through the HiCExperiment constructor function or using the import method implemented by HiCExperiment package.

HiCExperiment methods.

Usage

HiCExperiment(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(compartments = GenomicRanges::GRanges(),
    borders = GenomicRanges::GRanges(), loops =
    InteractionSet::GInteractions(GenomicRanges::GRanges(), GenomicRanges::GRanges()),
    viewpoints = GenomicRanges::GRanges()),
  pairsFile = NULL,
  ...
)

.HiCExperimentFromCoolFile(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(compartments = GenomicRanges::GRanges(),
    borders = GenomicRanges::GRanges(), loops =
    InteractionSet::GInteractions(GenomicRanges::GRanges(), GenomicRanges::GRanges()),
    viewpoints = GenomicRanges::GRanges()),
  pairsFile = NULL
)

.HiCExperimentFromHicFile(
  file,
  resolution = NULL,
  focus = NULL,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(compartments = GenomicRanges::GRanges(),
    borders = GenomicRanges::GRanges(), loops =
    InteractionSet::GInteractions(GenomicRanges::GRanges(), GenomicRanges::GRanges()),
    viewpoints = GenomicRanges::GRanges()),
  pairsFile = NULL
)

.HiCExperimentFromHicproFile(
  file,
  bed,
  metadata = list(),
  topologicalFeatures = S4Vectors::SimpleList(compartments = GenomicRanges::GRanges(),
    borders = GenomicRanges::GRanges(), loops =
    InteractionSet::GInteractions(GenomicRanges::GRanges(), GenomicRanges::GRanges()),
    viewpoints = GenomicRanges::GRanges()),
  pairsFile = NULL
)

# S4 method for HiCExperiment
resolutions(x)

# S4 method for HiCExperiment
resolution(x)

# S4 method for HiCExperiment
focus(x)

# S4 method for HiCExperiment,character
focus(x) <- value

# S4 method for HiCExperiment,numeric
zoom(x, resolution)

# S4 method for HiCExperiment,character
refocus(x, focus)

# S4 method for HiCExperiment,missing
scores(x)

# S4 method for HiCExperiment,character
scores(x, name)

# S4 method for HiCExperiment,numeric
scores(x, name)

# S4 method for HiCExperiment,character,numeric
scores(x, name) <- value

# S4 method for HiCExperiment,missing
topologicalFeatures(x)

# S4 method for HiCExperiment,character
topologicalFeatures(x, name)

# S4 method for HiCExperiment,numeric
topologicalFeatures(x, name)

# S4 method for HiCExperiment,character,GRangesOrGInteractions
topologicalFeatures(x, name) <- value

# S4 method for HiCExperiment
pairsFile(x)

# S4 method for HiCExperiment,character
pairsFile(x) <- value

# S4 method for HiCExperiment,list
metadata(x) <- value

# S4 method for HiCExperiment
fileName(object)

# S4 method for HiCExperiment
interactions(x)

# S4 method for HiCExperiment,GInteractions
interactions(x) <- value

# S4 method for HiCExperiment
length(x)

# S4 method for HiCExperiment,numeric,ANY,ANY
[(x, i)

# S4 method for HiCExperiment,logical,ANY,ANY
[(x, i)

# S4 method for HiCExperiment,character,ANY,ANY
[(x, i)

# S4 method for HiCExperiment
seqinfo(x)

# S4 method for HiCExperiment
bins(x)

# S4 method for HiCExperiment
anchors(x)

# S4 method for HiCExperiment
regions(x)

# S4 method for HiCExperiment
show(object)

# S4 method for HiCExperiment
as.matrix(x)

# S4 method for HiCExperiment
as.data.frame(x)

Arguments

file

CoolFile or plain path to a Hi-C contact file

resolution

Resolution to use with the Hi-C contact file

focus

Chromosome coordinates for which interaction counts are extracted from the Hi-C contact file, provided as a character string (e.g. "II:4000-5000"). If not provided, the entire Hi-C contact file will be imported.

metadata

list of metadata

topologicalFeatures

topologicalFeatures provided as a named SimpleList

pairsFile

Path to an associated .pairs file

...

Extra arguments

bed

Path to regions file generated by HiC-Pro

x

A HiCExperiment object.

value

...

name

...

object

A HiCExperiment object.

i

a range or boolean vector.

Value

An HiCExperiment object.

Slots

fileName

Path of Hi-C contact file

focus

Chr. coordinates for which interaction counts are extracted from the Hi-C contact file.

resolutions

Resolutions available in the Hi-C contact file.

resolution

Current resolution

interactions

Genomic Interactions extracted from the Hi-C contact file

scores

Available interaction scores.

topologicalFeatures

Topological features associated with the dataset (e.g. loops (\<Pairs\>), borders (\<GRanges\>), viewpoints (\<GRanges\>), etc...)

pairsFile

Path to the .pairs file associated with the Hi-C contact file

metadata

metadata associated with the Hi-C contact file.

Examples

mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
#> snapshotDate(): 2023-01-04
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
contacts_yeast <- HiCExperiment(mcool_path, resolution = 16000, focus = 'II')
contacts_yeast
#> `HiCExperiment` object with 1,308 interactions over 51 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/4d44b9a9af_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 16000 
#> interactions: 1308 
#> scores(2): count balanced 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(0):
resolutions(contacts_yeast)
#> [1]  1000  2000  4000  8000 16000
resolution(contacts_yeast)
#> [1] 16000
focus(contacts_yeast)
#> [1] "II"
scores(contacts_yeast)
#> List of length 2
#> names(2): count balanced
tail(scores(contacts_yeast, 1))
#> [1] 4797 3704  841 4696 2103 1040
tail(scores(contacts_yeast, 'balanced'))
#> [1] 0.6828679 0.7918215 0.3992624 1.5075570 1.4993084 1.6466114
scores(contacts_yeast, 'test') <- runif(length(contacts_yeast))
tail(scores(contacts_yeast, 'test'))
#> [1] 0.34144258 0.78136986 0.02337304 0.33026780 0.06262161 0.21015766
data(centros_yeast)
topologicalFeatures(contacts_yeast, 'centromeres') <- centros_yeast
topologicalFeatures(contacts_yeast, 1)
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
topologicalFeatures(contacts_yeast, 'centromeres')
#> GRanges object with 16 ranges and 0 metadata columns:
#>        seqnames        ranges strand
#>           <Rle>     <IRanges>  <Rle>
#>    [1]        I 151583-151641      +
#>    [2]       II 238361-238419      +
#>    [3]      III 114322-114380      +
#>    [4]       IV 449879-449937      +
#>    [5]        V 152522-152580      +
#>    ...      ...           ...    ...
#>   [12]      XII 151366-151424      +
#>   [13]     XIII 268222-268280      +
#>   [14]      XIV 628588-628646      +
#>   [15]       XV 326897-326955      +
#>   [16]      XVI 556255-556313      +
#>   -------
#>   seqinfo: 17 sequences (1 circular) from R64-1-1 genome
pairsFile(contacts_yeast)
#> NULL
fileName(contacts_yeast)
#> [1] "/github/home/.cache/R/ExperimentHub/4d44b9a9af_7752"
interactions(contacts_yeast)
#> GInteractions object with 1308 interactions and 5 metadata columns:
#>          seqnames1       ranges1     seqnames2       ranges2 |   bin_id1
#>              <Rle>     <IRanges>         <Rle>     <IRanges> | <integer>
#>      [1]        II       1-16000 ---        II       1-16000 |        15
#>      [2]        II       1-16000 ---        II   16001-32000 |        15
#>      [3]        II       1-16000 ---        II   32001-48000 |        15
#>      [4]        II       1-16000 ---        II   48001-64000 |        15
#>      [5]        II       1-16000 ---        II   64001-80000 |        15
#>      ...       ...           ... ...       ...           ... .       ...
#>   [1304]        II 768001-784000 ---        II 784001-800000 |        63
#>   [1305]        II 768001-784000 ---        II 800001-813184 |        63
#>   [1306]        II 784001-800000 ---        II 784001-800000 |        64
#>   [1307]        II 784001-800000 ---        II 800001-813184 |        64
#>   [1308]        II 800001-813184 ---        II 800001-813184 |        65
#>            bin_id2     count  balanced      test
#>          <integer> <numeric> <numeric> <numeric>
#>      [1]        15       990 1.1057215  0.903238
#>      [2]        16      2154 1.8846799  0.793748
#>      [3]        17       795 0.4339477  0.451694
#>      [4]        18       531 0.2091263  0.845529
#>      [5]        19       227 0.0950672  0.440947
#>      ...       ...       ...       ...       ...
#>   [1304]        64      3704  0.791822 0.7813699
#>   [1305]        65       841  0.399262 0.0233730
#>   [1306]        64      4696  1.507557 0.3302678
#>   [1307]        65      2103  1.499308 0.0626216
#>   [1308]        65      1040  1.646611 0.2101577
#>   -------
#>   regions: 51 ranges and 3 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
length(contacts_yeast)
#> [1] 1308
contacts_yeast[seq_len(10)]
#> `HiCExperiment` object with 10 interactions over 10 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/4d44b9a9af_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 16000 
#> interactions: 10 
#> scores(3): count balanced test 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) centromeres(16) 
#> pairsFile: N/A 
#> metadata(0):
seqinfo(contacts_yeast)
#> Seqinfo object with 16 sequences from an unspecified genome:
#>   seqnames seqlengths isCircular genome
#>   I            230218       <NA>   <NA>
#>   II           813184       <NA>   <NA>
#>   III          316620       <NA>   <NA>
#>   IV          1531933       <NA>   <NA>
#>   V            576874       <NA>   <NA>
#>   ...             ...        ...    ...
#>   XII         1078177       <NA>   <NA>
#>   XIII         924431       <NA>   <NA>
#>   XIV          784333       <NA>   <NA>
#>   XV          1091291       <NA>   <NA>
#>   XVI          948066       <NA>   <NA>
bins(contacts_yeast)
#> GRanges object with 763 ranges and 2 metadata columns:
#>                     seqnames        ranges strand |    bin_id     weight
#>                        <Rle>     <IRanges>  <Rle> | <numeric>  <numeric>
#>           I_1_16000        I       1-16000      * |         0  0.0196442
#>       I_16001_32000        I   16001-32000      * |         1  0.0220746
#>       I_32001_48000        I   32001-48000      * |         2  0.0188701
#>       I_48001_64000        I   48001-64000      * |         3  0.0136679
#>       I_64001_80000        I   64001-80000      * |         4  0.0134860
#>                 ...      ...           ...    ... .       ...        ...
#>   XVI_880001_896000      XVI 880001-896000      * |       758 0.00910873
#>   XVI_896001_912000      XVI 896001-912000      * |       759 0.01421350
#>   XVI_912001_928000      XVI 912001-928000      * |       760 0.02439992
#>   XVI_928001_944000      XVI 928001-944000      * |       761 0.01993237
#>   XVI_944001_948066      XVI 944001-948066      * |       762        NaN
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
anchors(contacts_yeast)
#> $first
#> GRanges object with 1308 ranges and 3 metadata columns:
#>          seqnames        ranges strand |   chr    center    bin_id
#>             <Rle>     <IRanges>  <Rle> | <Rle> <integer> <numeric>
#>      [1]       II       1-16000      * |    II      8000        15
#>      [2]       II       1-16000      * |    II      8000        15
#>      [3]       II       1-16000      * |    II      8000        15
#>      [4]       II       1-16000      * |    II      8000        15
#>      [5]       II       1-16000      * |    II      8000        15
#>      ...      ...           ...    ... .   ...       ...       ...
#>   [1304]       II 768001-784000      * |    II    776000        63
#>   [1305]       II 768001-784000      * |    II    776000        63
#>   [1306]       II 784001-800000      * |    II    792000        64
#>   [1307]       II 784001-800000      * |    II    792000        64
#>   [1308]       II 800001-813184      * |    II    806592        65
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
#> $second
#> GRanges object with 1308 ranges and 3 metadata columns:
#>          seqnames        ranges strand |   chr    center    bin_id
#>             <Rle>     <IRanges>  <Rle> | <Rle> <integer> <numeric>
#>      [1]       II       1-16000      * |    II      8000        15
#>      [2]       II   16001-32000      * |    II     24000        16
#>      [3]       II   32001-48000      * |    II     40000        17
#>      [4]       II   48001-64000      * |    II     56000        18
#>      [5]       II   64001-80000      * |    II     72000        19
#>      ...      ...           ...    ... .   ...       ...       ...
#>   [1304]       II 784001-800000      * |    II    792000        64
#>   [1305]       II 800001-813184      * |    II    806592        65
#>   [1306]       II 784001-800000      * |    II    792000        64
#>   [1307]       II 800001-813184      * |    II    806592        65
#>   [1308]       II 800001-813184      * |    II    806592        65
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
#> 
regions(contacts_yeast)
#> GRanges object with 51 ranges and 3 metadata columns:
#>                    seqnames        ranges strand |   chr    center    bin_id
#>                       <Rle>     <IRanges>  <Rle> | <Rle> <integer> <numeric>
#>         II_1_16000       II       1-16000      * |    II      8000        15
#>     II_16001_32000       II   16001-32000      * |    II     24000        16
#>     II_32001_48000       II   32001-48000      * |    II     40000        17
#>     II_48001_64000       II   48001-64000      * |    II     56000        18
#>     II_64001_80000       II   64001-80000      * |    II     72000        19
#>                ...      ...           ...    ... .   ...       ...       ...
#>   II_736001_752000       II 736001-752000      * |    II    744000        61
#>   II_752001_768000       II 752001-768000      * |    II    760000        62
#>   II_768001_784000       II 768001-784000      * |    II    776000        63
#>   II_784001_800000       II 784001-800000      * |    II    792000        64
#>   II_800001_813184       II 800001-813184      * |    II    806592        65
#>   -------
#>   seqinfo: 16 sequences from an unspecified genome
contacts_yeast
#> `HiCExperiment` object with 1,308 interactions over 51 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/4d44b9a9af_7752" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 16000 
#> interactions: 1308 
#> scores(3): count balanced test 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) centromeres(16) 
#> pairsFile: N/A 
#> metadata(0):
as(contacts_yeast, 'GInteractions')
#> GInteractions object with 1308 interactions and 5 metadata columns:
#>          seqnames1       ranges1     seqnames2       ranges2 |   bin_id1
#>              <Rle>     <IRanges>         <Rle>     <IRanges> | <integer>
#>      [1]        II       1-16000 ---        II       1-16000 |        15
#>      [2]        II       1-16000 ---        II   16001-32000 |        15
#>      [3]        II       1-16000 ---        II   32001-48000 |        15
#>      [4]        II       1-16000 ---        II   48001-64000 |        15
#>      [5]        II       1-16000 ---        II   64001-80000 |        15
#>      ...       ...           ... ...       ...           ... .       ...
#>   [1304]        II 768001-784000 ---        II 784001-800000 |        63
#>   [1305]        II 768001-784000 ---        II 800001-813184 |        63
#>   [1306]        II 784001-800000 ---        II 784001-800000 |        64
#>   [1307]        II 784001-800000 ---        II 800001-813184 |        64
#>   [1308]        II 800001-813184 ---        II 800001-813184 |        65
#>            bin_id2     count  balanced      test
#>          <integer> <numeric> <numeric> <numeric>
#>      [1]        15       990 1.1057215  0.903238
#>      [2]        16      2154 1.8846799  0.793748
#>      [3]        17       795 0.4339477  0.451694
#>      [4]        18       531 0.2091263  0.845529
#>      [5]        19       227 0.0950672  0.440947
#>      ...       ...       ...       ...       ...
#>   [1304]        64      3704  0.791822 0.7813699
#>   [1305]        65       841  0.399262 0.0233730
#>   [1306]        64      4696  1.507557 0.3302678
#>   [1307]        65      2103  1.499308 0.0626216
#>   [1308]        65      1040  1.646611 0.2101577
#>   -------
#>   regions: 51 ranges and 3 metadata columns
#>   seqinfo: 16 sequences from an unspecified genome
as(contacts_yeast, 'ContactMatrix')
#> class: ContactMatrix 
#> dim: 51 51 
#> type: dgCMatrix 
#> rownames: NULL
#> colnames: NULL
#> metadata(0):
#> regions: 51
as(contacts_yeast, 'matrix')[seq_len(10), seq_len(10)]
#>              [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 1.105721482 1.88467986 0.43394771 0.20912625 0.09506719 0.04461248
#>  [2,] 1.884679865 2.69106090 0.77184273 0.31223090 0.16207391 0.09332190
#>  [3,] 0.433947710 0.77184273 0.88539537 0.49138803 0.19485048 0.10066488
#>  [4,] 0.209126253 0.31223090 0.49138803 0.70089155 0.52173737 0.18074062
#>  [5,] 0.095067195 0.16207391 0.19485048 0.52173737 0.80119743 0.55114262
#>  [6,] 0.044612481 0.09332190 0.10066488 0.18074062 0.55114262 0.80971473
#>  [7,] 0.034479507 0.06966022 0.05454341 0.08878588 0.21060194 0.64080581
#>  [8,] 0.021145336 0.03438054 0.02885760 0.04361179 0.09993312 0.22175041
#>  [9,] 0.013121683 0.02925696 0.02121164 0.02669383 0.05185146 0.09843438
#> [10,] 0.009843827 0.01606586 0.01383125 0.01402904 0.02906758 0.05333252
#>             [,7]       [,8]       [,9]       [,10]
#>  [1,] 0.03447951 0.02114534 0.01312168 0.009843827
#>  [2,] 0.06966022 0.03438054 0.02925696 0.016065864
#>  [3,] 0.05454341 0.02885760 0.02121164 0.013831254
#>  [4,] 0.08878588 0.04361179 0.02669383 0.014029040
#>  [5,] 0.21060194 0.09993312 0.05185146 0.029067576
#>  [6,] 0.64080581 0.22175041 0.09843438 0.053332524
#>  [7,] 0.83295742 0.54630400 0.19556965 0.093798964
#>  [8,] 0.54630400 0.59301354 0.58806903 0.241309551
#>  [9,] 0.19556965 0.58806903 0.74046424 0.524313830
#> [10,] 0.09379896 0.24130955 0.52431383 0.890641254
as(contacts_yeast, 'data.frame')[seq_len(10), seq_len(10)]
#>    seqnames1 start1  end1 width1 strand1 center1 bin_id1 seqnames2 start2
#> 1         II      1 16000  16000       *    8000      15        II      1
#> 2         II      1 16000  16000       *    8000      15        II  16001
#> 3         II      1 16000  16000       *    8000      15        II  32001
#> 4         II      1 16000  16000       *    8000      15        II  48001
#> 5         II      1 16000  16000       *    8000      15        II  64001
#> 6         II      1 16000  16000       *    8000      15        II  80001
#> 7         II      1 16000  16000       *    8000      15        II  96001
#> 8         II      1 16000  16000       *    8000      15        II 112001
#> 9         II      1 16000  16000       *    8000      15        II 128001
#> 10        II      1 16000  16000       *    8000      15        II 144001
#>      end2
#> 1   16000
#> 2   32000
#> 3   48000
#> 4   64000
#> 5   80000
#> 6   96000
#> 7  112000
#> 8  128000
#> 9  144000
#> 10 160000