HicproFile S4 class

The HicproFile class describes a BiocFile object, pointing to the location of a HiC-Pro-generated matrix file and containing 4 additional slots:

bed: path to the matching .bed file generated by HiC-Pro;

resolution: at which resolution the associated mcool file should be parsed ;

pairsFile: the path (in plain character) to an optional pairs file (stored as a PairsFile object);

metadata: a list metadata

HicproFile methods.

Arguments

path

String; path to the HiC-Pro output .matrix file (matrix file)

bed

String; path to the HiC-Pro output .bed file (regions file)

pairsFile

String; path to a pairs file

metadata

list.

object

A HicproFile object.

Slots

bed

Path to the matching .bed file generated by HiC-Pro

See also

Examples

hicproMatrixPath <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_matrix')
#> snapshotDate(): 2023-03-29
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
hicproBedPath <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_bed')
#> snapshotDate(): 2023-03-29
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
pairsPath <- HiContactsData::HiContactsData('yeast_wt', 'pairs.gz')
#> snapshotDate(): 2023-03-29
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
hicpro <- HicproFile(
  hicproMatrixPath, bed = hicproBedPath, pairs = pairsPath ,
  metadata = list(type = 'example')
)
hicpro
#> HicproFile object
#> HiC-Pro files:
#>   $ matrix:   /github/home/.cache/R/ExperimentHub/49d57e2ee8d_7837 
#>   $ regions:  /github/home/.cache/R/ExperimentHub/49d1978d3fd_7838 
#> resolution: 1000 
#> pairs file: /github/home/.cache/R/ExperimentHub/49d70352d5a_7753 
#> metadata(1): type
resolution(hicpro)
#> [1] 1000
pairsFile(hicpro)
#>                                                 EH7703 
#> "/github/home/.cache/R/ExperimentHub/49d70352d5a_7753" 
metadata(hicpro)
#> $type
#> [1] "example"
#>