HiCExperiment import methods

Import methods for data structures implemented in the HiCExperiment package. HiCExperiment package implements methods to faciliate the import of Hi-C files (.(m)cool, .hic, HiC-Pro derived) and pairs files in R, as HiCExperiment or GInteractions objects.

import(con, format, text, ...)

Arguments

con

Path or connection to a cool, mcool, .hic or HiC-Pro derived files. Can also be a path to a pairs file.

format

The format of the output. If missing and 'con' is a filename, the format is derived from the file extension. This argument is unnecessary when files are directly provided as CoolFile, HicFile, HicproFile or PairsFile.

text

If 'con' is missing, this can be a character vector directly providing the string data to import.

...

e.g. resolution = ...; parameters to pass to format-specific methods.

Value

A HiCExperiment or GInteractions object

Examples

# ---- Importing .(m)cool files 
mcool_path <- HiContactsData::HiContactsData('yeast_wt', 'mcool')
#> snapshotDate(): 2022-12-16
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
import(mcool_path, resolution = 16000, format = 'cool')
#> `HiCExperiment` object with 267,709 interactions over 759 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/108c221286366_7752" 
#> focus: "whole genome" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 16000 
#> interactions: 267709 
#> scores(2): count balanced 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(0):

# ---- Importing .hic files 
hic_path <- HiContactsData::HiContactsData('yeast_wt', 'hic')
#> snapshotDate(): 2022-12-16
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
import(hic_path, resolution = 2000, focus = 'II', format = 'hic')
#> `HiCExperiment` object with 34,179 interactions over 404 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/108c237603bc1_7836" 
#> focus: "II" 
#> resolutions(5): 1000 2000 4000 8000 16000
#> current resolution: 2000 
#> interactions: 34179 
#> scores(2): count balanced 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(0):

# ---- Importing HiC-Pro files 
hicpro_matrix_path <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_matrix')
#> snapshotDate(): 2022-12-16
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
hicpro_bed_path <- HiContactsData::HiContactsData('yeast_wt', 'hicpro_bed')
#> snapshotDate(): 2022-12-16
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
import(hicpro_matrix_path, bed = hicpro_bed_path, format = 'hicpro')
#> `HiCExperiment` object with 2,686,250 interactions over 11,805 regions 
#> -------
#> fileName: "/github/home/.cache/R/ExperimentHub/108c26c7e0c4e_7837" 
#> focus: "whole genome" 
#> resolutions(1): 1000
#> current resolution: 1000 
#> interactions: 2686250 
#> scores(1): count 
#> topologicalFeatures: compartments(0) borders(0) loops(0) viewpoints(0) 
#> pairsFile: N/A 
#> metadata(1): regions